.local/src/spotify-release-radar-save/pyproject.toml
[tool.poetry]
name = "spotify-release-radar-save"
version = "0.1.0"
description = ""
authors = ["Aleksandr Verevkin <sandyverevkin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
spotipy = "^2.23.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
